requirements.txt
django-jazzmin==2.6.0
django-import-export==3.3.7
environs==10.3.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
djangorestframework~=3.14.0
Pillow-10.2.0